+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to XML.jl will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.3] - 2026-07-26
 
 ### Added
 
@@ -310,7 +310,10 @@ First release since XML.jl moved to [JuliaData](https://github.com/JuliaData/XML
 
 - Initial release.
 
-[Unreleased]: https://github.com/JuliaData/XML.jl/compare/v0.3.9...HEAD
+[0.4.3]: https://github.com/JuliaData/XML.jl/compare/v0.4.2...v0.4.3
+[0.4.2]: https://github.com/JuliaData/XML.jl/compare/v0.4.1...v0.4.2
+[0.4.1]: https://github.com/JuliaData/XML.jl/compare/v0.4.0...v0.4.1
+[0.4.0]: https://github.com/JuliaData/XML.jl/compare/v0.3.9...v0.4.0
 [0.3.9]: https://github.com/JuliaData/XML.jl/compare/v0.3.8...v0.3.9
 [0.3.8]: https://github.com/JuliaData/XML.jl/compare/v0.3.7...v0.3.8
 [0.3.7]: https://github.com/JuliaData/XML.jl/compare/v0.3.6...v0.3.7